+++ wargcore/app_cli.go
@@ -221,7 +221,7 @@ func (app *App) Validate() error {
 	return nil
 }
 
-type CompletionCandidates func(Context) (*completion.Candidates, error)
+type CompletionCandidatesFunc func(Context) (*completion.Candidates, error)
 
 func (a *App) CompletionCandidates(opts ...ParseOpt) (*completion.Candidates, error) {
 	parseOpts := NewParseOpts(opts...)