+++ wargcore/command_cli.go
@@ -54,7 +54,7 @@ type Command struct {
 	// CompletionCandidates is a function that returns a list of completion candidates for this commmand.
 	// Note that some flags in the cli.Context Flags map may not be set, even if they're required.
 	// TODO: get a comprehensive list of restrictions on the context.
-	CompletionCandidates CompletionCandidates
+	CompletionCandidates CompletionCandidatesFunc
 
 	// Parsed Flags
 	Flags FlagMap