+++ wargcore/flag_cli.go
@@ -44,7 +44,7 @@ type Flag struct {
 	// CompletionCandidates is a function that returns a list of completion candidates for this flag.
 	// Note that some flags in the cli.Context Flags map may not be set, even if they're required.
 	// TODO: get a comprehensive list of restrictions on the context.
-	CompletionCandidates CompletionCandidates
+	CompletionCandidates CompletionCandidatesFunc
 
 	// ConfigPath is the path from the config to the value the flag updates
 	ConfigPath string